requirements.txt
requests
qrcode
pillow
pyserial
tk